backend/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alembic==1.14.1
annotated-types==0.7.0
anthropic==0.45.2
anyio==4.8.0
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
defusedxml==0.7.1
Deprecated==1.2.15
distro==1.9.0
fastapi==0.115.6
frozenlist==1.5.0
gitdb==4.0.12
gitingest==0.1.2
GitPython==3.1.44
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.15
langchain-anthropic==0.3.4
langchain-core==0.3.31
langchain-openai==0.3.2
langchain-text-splitters==0.3.5
langsmith==0.3.1
Mako==1.3.8
MarkupSafe==3.0.2
multidict==6.1.0
numpy==2.2.2
openai==1.60.1
orjson==3.10.15
packaging==24.2
pinecone==5.4.2
pinecone-plugin-inference==3.1.0
pinecone-plugin-interface==0.0.7
propcache==0.2.1
psycopg2-binary==2.9.10
pydantic==2.10.5
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.41.3
svix==1.56.0
tenacity==9.0.0
tiktoken==0.8.0
tqdm==4.67.1
types-Deprecated==1.2.15.20241117
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
wrapt==1.17.2
yarl==1.18.3
zstandard==0.23.0
